fix router and handler
