--- internal/user/user_handler.go
@@ -1,99 +1,0 @@
-package user
-
-import (
-	"errors"
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-)
-
-// maneja el Service de User e implementa handlers http
-type handler struct {
-	userService *Service
-}
-
-// handleCreate maneja POST /users
-func (h *handler) handleCreate(ctx *gin.Context) {
-	// request payload = datos
-	var req struct {
-		Name     string `json:"name"`
-		Address  string `json:"address"`
-		NickName string `json:"nickname"`
-	}
-	if err := ctx.ShouldBindJSON(&req); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	u := &User{
-		Name:     req.Name,
-		Address:  req.Address,
-		NickName: req.NickName,
-	}
-	if err := h.userService.Create(u); err != nil {
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	ctx.JSON(http.StatusCreated, u)
-}
-
-// handleRead maneja GET /users/:id
-func (h *handler) handleRead(ctx *gin.Context) {
-	id := ctx.Param("id")
-
-	u, err := h.userService.Get(id)
-	if err != nil {
-		if errors.Is(err, ErrNotFound) {
-			ctx.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
-			return
-		}
-
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	ctx.JSON(http.StatusOK, u)
-}
-
-// handleUpdate maneja PUT /users/:id
-func (h *handler) handleUpdate(ctx *gin.Context) {
-	id := ctx.Param("id")
-
-	// bind partial update fields
-	var fields *UpdateFields
-	if err := ctx.ShouldBindJSON(&fields); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	u, err := h.userService.Update(id, fields)
-	if err != nil {
-		if errors.Is(err, ErrNotFound) {
-			ctx.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
-			return
-		}
-
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	ctx.JSON(http.StatusOK, u)
-}
-
-// handleDelete maneja DELETE /users/:id
-func (h *handler) handleDelete(ctx *gin.Context) {
-	id := ctx.Param("id")
-
-	if err := h.userService.Delete(id); err != nil {
-		if errors.Is(err, ErrNotFound) {
-			ctx.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
-			return
-		}
-
-		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
-		return
-	}
-
-	ctx.Status(http.StatusNoContent)
-}
